fix: Modificacion del valor de un campo de edicion

+++ kotlinsql/src/main/kotlin/com/example/kotlinsql/model/edicion/Edicion.kt
@@ -24,7 +24,7 @@ data class Edicion(
 
     @ManyToOne
     @JoinColumn(name = "curso_id", nullable = true)
-    val curso: Int? = null,
+    val curso: Long? = null,
 
     val estado: Boolean = true
 )
